LPS-67714 Refactor MB attachments to be migrated following the same pattern as Wiki attachments (with a DLStoreConvertProcess)
